backend/requirements.txt
fastapi
uvicorn
boto3
azure-identity
azure-mgmt-costmanagement
starlette
